Edits for clarity on Lesson 3 ROS Nodes
Felt as though it's pretty crucial to tell them to use a newer version of catkin commands than is listed in the tutorial, so it's not something we should make them go digging for. Hence why I moved all of the relevant info and made it more explicit in the Goals section, despite trying to usually keep it short.